Dispatcher

See merge request lobnews/ComponentwarePraktikum2!27

+++ FortDay-Dispatcher-ejb/ejbModule/de/fh_dortmund/inf/cw/surstwalat/dispatcher/interfaces/GameRepositoryLocal.java
@@ -2,9 +2,13 @@ package de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces;
 
 import javax.ejb.Local;
 
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.domain.Game;
-
+import de.fh_dortmund.inf.cw.surstwalat.common.model.Game;
+/**
+ * 
+ * @author Johannes Heiderich
+ *
+ */
 @Local
-public interface GameRepositoryLocal extends IDispatcherRepository<Game, Long> {
+public interface GameRepositoryLocal extends IDispatcherRepository<Game, Integer> {
 
 }
