+++ FortDay-Dispatcher-ejb/ejbModule/de/fh_dortmund/inf/cw/surstwalat/dispatcher/interfaces/PlayerRepositoryLocal.java
@@ -2,9 +2,13 @@ package de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces;
 
 import javax.ejb.Local;
 
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.domain.Player;
-
+import de.fh_dortmund.inf.cw.surstwalat.common.model.Player;
+/**
+ * 
+ * @author Johannes Heiderich
+ *
+ */
 @Local
-public interface PlayerRepositoryLocal extends IDispatcherRepository<Player, Long> {
+public interface PlayerRepositoryLocal extends IDispatcherRepository<Player, Integer> {
 
 }
